VolumeControl moved to its own file, code cleaned up, README updated
