+++ package.json
@@ -29,7 +29,6 @@
     "@react-navigation/native": "^6.0.2",
     "@react-navigation/native-stack": "^6.1.0",
     "@reduxjs/toolkit": "1.8.0",
-    "@sharcoux/slider": "5.6.5",
     "date-fns": "2.28.0",
     "date-fns-tz": "1.3.1",
     "expo": "~44.0.0",
